Moved common data models, data stores, and database context to a seperate project to act as a base context for both server and client databases; to support caching in client machine.
