+++ Source/InvestorAPI/InvestorAPI.Data/DataStores/Abstraction/Prototype/IBrandStore.cs
@@ -1,4 +1,6 @@
-﻿namespace InvestorAPI.Data
+﻿using InvestorData;
+
+namespace InvestorAPI.Data
 {
 
     /// <summary>
