--- Source/InvestorAPI/InvestorAPI.Data/DataStores/Base/Enums.cs
@@ -1,16 +1,0 @@
-﻿namespace InvestorAPI.Data
-{
-    public enum LoginResult
-    {
-        Success,
-        Failed,
-        UserNotFound
-    }
-    
-    public enum DeleteResult
-    {
-        Success,
-        Failed,
-        EntityNotFound
-    }
-}
