+++ Source/InvestorAPI/InvestorAPI.Data/InvestorAPI.Data.csproj
@@ -19,4 +19,8 @@
     <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="6.0.8" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\InvestorData\InvestorData\InvestorData.csproj" />
+  </ItemGroup>
+
 </Project>
