--- Source/InvestorAPI/InvestorAPI.Data/Models/Base/DatedEntity.cs
@@ -1,19 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace InvestorAPI.Data
-{
-    public abstract class DatedEntity
-    {
-
-        [Required]
-        [Precision(3)]
-        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public DateTime DateCreated { get; set; }
-        
-        [Precision(3)]
-        public DateTime? DateModified { get; set; }
-
-    }
-}
