--- Source/InvestorAPI/InvestorAPI.Data/Models/Contacts/Contact.cs
@@ -1,28 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace InvestorAPI.Data
-{
-    public class Contact : DatedEntity
-    {
-
-        [Key]
-        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public string Id { get; set; } = null!;
-
-
-        [MaxLength(64)]
-        public string? FirstName { get; set; }
-
-        [MaxLength(64)]
-        public string? LastName { get; set; }
-
-
-        [MaxLength(128)]
-        public string? Email { get; set; }
-
-        [MaxLength(32)]
-        public string? Phone { get; set; }
-
-    }
-}
