--- Source/InvestorAPI/InvestorAPI.Data/Models/Management/BusinessType.cs
@@ -1,22 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace InvestorAPI.Data
-{
-    public class BusinessType 
-    {
-
-        [Key]
-        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public string Id { get; set; } = null!;
-
-
-        [Required]
-        [MaxLength(256)]
-        public string Name { get; set; } = null!;
-
-        [MaxLength(1024)]
-        public string? Description { get; set; }
-
-    }
-}
