--- Source/InvestorAPI/InvestorAPI.Data/Models/Products/Category.cs
@@ -1,28 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace InvestorAPI.Data
-{
-    public class Category : DatedEntity
-    {
-
-        [Key]
-        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public string Id { get; set; } = null!;
-
-        public Business? Business { get; set; }
-        public string? BusinessId { get; set; }
-
-
-        [Required]
-        [MaxLength(256)]
-        public string Name { get; set; } = null!;
-
-        [MaxLength(1024)]
-        public string? Description { get; set; }
-
-
-        public List<Product> Products { get; set; } = new();
-
-    }
-}
