+++ Source/InvestorData/InvestorData.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.3.32825.248
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InvestorData", "InvestorData\InvestorData.csproj", "{6397E4EE-7766-47C2-9542-49ABC942E5AB}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6397E4EE-7766-47C2-9542-49ABC942E5AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6397E4EE-7766-47C2-9542-49ABC942E5AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6397E4EE-7766-47C2-9542-49ABC942E5AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6397E4EE-7766-47C2-9542-49ABC942E5AB}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {ABB3DEAC-6035-42C6-A3C1-9F9CA1149B2E}
+	EndGlobalSection
+EndGlobal
