+++ Source/InvestorData/InvestorData/DataStores/Base/Enums.cs
@@ -1,0 +1,16 @@
+﻿namespace InvestorData
+{
+    public enum LoginResult
+    {
+        Success,
+        Failed,
+        UserNotFound
+    }
+    
+    public enum DeleteResult
+    {
+        Success,
+        Failed,
+        EntityNotFound
+    }
+}
