+++ Source/InvestorData/InvestorData/Models/Accounting/Invoice.cs
@@ -1,0 +1,28 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.Diagnostics.CodeAnalysis;
+
+namespace InvestorData
+{
+    public class Invoice : InvoiceBillBase
+    {
+
+        [Key]
+        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        public string Id { get; set; } = null!;
+
+        [Required]
+        public string BusinessId { get; set; } = null!;
+        public Business Business { get; set; } = null!;
+
+
+        [NotNull]
+        public Customer? Customer { get; set; }
+        public string? CustomerId { get; set; }
+
+
+        public List<InvoiceItem> Items { get; set; } = new();
+
+    }
+
+}
