+++ Source/InvestorData/InvestorData/Models/Accounting/PaymentMethod.cs
@@ -1,0 +1,25 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace InvestorData
+{
+    public class PaymentMethod
+    {
+
+        [Key]
+        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        public string Id { get; set; } = null!;
+
+        public Business? Business { get; set; }
+        public string? BusinessId { get; set; }
+
+
+        [Required]
+        [MaxLength(256)]
+        public string Name { get; set; } = null!;
+
+        [MaxLength(1024)]
+        public string? Description { get; set; }
+
+    }
+}
