+++ Source/InvestorData/InvestorData/Models/Base/DatedEntity.cs
@@ -1,0 +1,19 @@
+﻿using Microsoft.EntityFrameworkCore;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace InvestorData
+{
+    public abstract class DatedEntity
+    {
+
+        [Required]
+        [Precision(3)]
+        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        public DateTime DateCreated { get; set; }
+        
+        [Precision(3)]
+        public DateTime? DateModified { get; set; }
+
+    }
+}
