+++ Source/InvestorData/InvestorData/Models/Base/PaymentBase.cs
@@ -1,0 +1,24 @@
+﻿using Microsoft.EntityFrameworkCore;
+using System.ComponentModel.DataAnnotations;
+
+namespace InvestorData
+{
+    public abstract class PaymentBase : DatedEntity
+    {
+
+        public int? Number { get; set; }
+
+        [Required]
+        public decimal Amount { get; set; }
+
+
+        [Required]
+        [Precision(0)]
+        public DateTime PaymentDate { get; set; }
+
+
+        [MaxLength(1024)]
+        public string? Notes { get; set; }
+
+    }
+}
